backup: save Tmux session in which I rolled back a Homebrew cask application (KeePassXC)
Raycast's KeePassXC plugin depends on the `keepassxc-cli`, so this
rollback preservers compatibility.

KeePassXC version 2.7.11 has issues with some binaries that it ships
with: namely `keepassxc-cli`, but also `keepassxc-proxy` (which is a
backend service. Running any `keepassxc-cli` command with 2.7.11
produces `Killed: 9`. For example:

$ keepassxc-cli --help
Killed: 9                  keepassxc-cli --help

To work around this, I rolled back to version 2.7.10.

References:
  - https://github.com/raycast/extensions/issues/23107
  - https://github.com/keepassxreboot/keepassxc/releases/tag/2.7.11
  - https://github.com/keepassxreboot/keepassxc/issues/12713#issuecomment-3570259643